Rewrite README to use Git LFS-hosted ES snapshot
The snapshot was being stored on Google Drive, which is not accessible
to external viewers. I therefore uploaded it to GitHub as multiple Git
LFS files. I described the process in the commit that added them.[^1]

I have updated the relevant README section to use the new zip files.
The provided commands should work on any Unix system.

[^1]: https://github.com/ALLAN-DIP/large-file-storage/commit/178bd645ffd28ae601b2502600943a20459b63f7